feat: split mealplangenerator-UI logic+state mgmt
