+++ client/src/components/meal_planning/MealPlanGenerator.tsx
@@ -1,101 +1,29 @@
-/* eslint-disable no-mixed-spaces-and-tabs */
-import { useState } from "react";
-import { useUserContext } from "../../hooks/useUserContext";
+import React from "react";
+import useSaveMealPlan from "../../hooks/useSaveMealPlan";
+import MealPlanCards from "../meal_planning/MealPlanCards";
+import NutrientsCard from "../meal_planning/NutrientsCard";
 import { MealPlanResponse } from "../../types/mealplan_response.type";
 
 interface MealPlanGeneratorProps {
   mealResponse: MealPlanResponse;
 }
 
-export const MealPlanGenerator: React.FC<MealPlanGeneratorProps> = ({
+const MealPlanGenerator: React.FC<MealPlanGeneratorProps> = ({
   mealResponse,
 }) => {
-  const [meals] = useState(mealResponse.meals);
-  const [nutrients] = useState(mealResponse.nutrients);
-  const [error, setError] = useState<string | null>(null);
-  const [successMessage, setSuccessMessage] = useState("");
-  const { id } = useUserContext();
-
-  const saveMealPlan = async (mealResponse: MealPlanResponse) => {
-    try {
-      const response = await fetch("https://localhost:3000/api/mealplan", {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify({
-          userId: id,
-          meals: mealResponse.meals,
-          nutrients: mealResponse.nutrients,
-        }),
-      });
-
-      if (response.ok) {
-        const data = await response.json();
-        if (data && data.message) {
-          setSuccessMessage(data.message);
-        } else {
-          setSuccessMessage("Meal plan saved successfully!");
-        }
-      } else {
-        const data = await response.json();
-        if (data && data.details) {
-          const errorMessage = data.details
-            .map((detail: { message: string }) => detail.message)
-            .join(" ");
-          throw new Error(errorMessage);
-        }
-        throw new Error(response.statusText || "Failed to save meal plan.");
-      }
-    } catch (err) {
-      if (err instanceof Error) {
-        setError(err.message);
-      } else {
-        setError("An unknown error occurred.");
-      }
-    }
-  };
+  const { meals, nutrients } = mealResponse;
+  const { saveMealPlan, error, successMessage } = useSaveMealPlan();
 
   return (
     <div className="bg-yellow-100">
       <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-4">
         {meals.map((meal) => (
-          <div
-            key={meal.id}
-            className="prominent-shadow rounded overflow:auto shadow-md bg-yellow-200 p-4 border-4 border-orange-400 h-96"
-          >
-            <img
-              src={`https://spoonacular.com/recipeImages/${meal.id}-240x150.${meal.imageType}`}
-              alt={meal.title}
-              className="object-cover h-48"
-            />
-            <div className="p-4">
-              <h2 className="font-bold text-xl mb-2">{meal.title}</h2>
-              <p>Ready In: {meal.readyInMinutes} minutes</p>
-              <p>Servings: {meal.servings}</p>
-              <p>
-                <a
-                  href={meal.sourceUrl}
-                  target="_blank"
-                  rel="noopener noreferrer"
-                  className="text-blue-500 hover:underline font bold"
-                >
-                  See Recipe
-                </a>
-              </p>
-            </div>
-          </div>
+          <MealPlanCards key={meal.id} meal={meal} />
         ))}
       </div>
 
       <div className="flex mt-4 items-center gap-2">
-        <div className="border-4  bg-yellow-200 border-orange-400 w-64 prominent-shadow rounded-md">
-          <h3 className="font-bold text-green-600">Nutritional Information:</h3>
-          <p>Calories: {nutrients?.calories}</p>
-          <p>Protein: {nutrients?.protein}g</p>
-          <p>Fat: {nutrients?.fat}g</p>
-          <p>Carbohydrates: {nutrients?.carbohydrates}g</p>
-        </div>
+        <NutrientsCard nutrients={nutrients} />
         <div>
           <button
             onClick={() => saveMealPlan(mealResponse)}
@@ -105,6 +33,7 @@ export const MealPlanGenerator: React.FC<MealPlanGeneratorProps> = ({
           </button>
         </div>
       </div>
+
       {successMessage && (
         <div className="text-green-500 mt-4 font-bold">{successMessage}</div>
       )}
