+++ client/src/types/mealplan_response.type.ts
@@ -1,14 +1,14 @@
-import { Meal } from './meal_data.type';
-import { Nutrients } from './nutrients_data.type';
+import { Meal } from "./meal_data.type";
+import { Nutrients } from "./nutrients_data.type";
 
 export type MealPlanResponse = {
-	status: number;
-	meals: Meal[];
-	nutrients: Nutrients;
+  status: number;
+  meals: Meal[];
+  nutrients: Nutrients;
 };
 
 export type MealPlan = {
-	userId: string;
-	meals: Meal[];
-	nutrients: Nutrients;
-}
+  userId: string;
+  meals: Meal[];
+  nutrients: Nutrients;
+};
